SUS-5481 | extract anonymous inner class
